feat(ride): insere a capacidade do veículo pra validações posteriores

+++ src/main/java/com/ifride/core/driver/model/dto/VehicleCreationDTO.java
@@ -2,5 +2,5 @@ package com.ifride.core.driver.model.dto;
 
 import java.math.BigInteger;
 
-public record VehicleCreationDTO(String model, String plate, String color, BigInteger capacity) {
+public record VehicleCreationDTO(String model, String plate, String color, Integer capacity) {
 }
